11/14 게시글 삭제시 권한 확인 (401,403)

+++ src/main/java/com/example/prj1be/controller/BoardController.java
@@ -50,7 +50,16 @@ public class BoardController {
     }
 
     @DeleteMapping("remove/{id}")
-    public ResponseEntity remove(@PathVariable Integer id) {
+    public ResponseEntity remove(@PathVariable Integer id,
+                                 @SessionAttribute(value = "login",required = false)Member login) {
+        if(login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); // 401 로그인을 안해서 권한이 없는 에러
+        }
+
+        if(!service.hasAccess(id, login)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 403 에러 로그인은 했지만 권한이 없는 에러
+        }
+
         if(service.remove(id)) {
             return ResponseEntity.ok().build();
         }else {
